move repository method location
